SignupRequest id, password, name, email
