+++ src/main/java/org/example/springdb/tiggle/user/model/request/SignupRequest.java
@@ -1,0 +1,45 @@
+package org.example.springdb.tiggle.user.model.request;
+
+public class SignupRequest {
+    private String id;
+    private String password;
+    private String name;
+    private String email;
+
+
+    public SignupRequest(String id, String password, String name, String email) {
+        this.id = id;
+        this.password = password;
+        this.name = name;
+        this.email = email;
+    }
+
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
