+++ src/main/java/org/example/springdb/tiggle/user/model/response/SignupResponse.java
@@ -1,0 +1,5 @@
+package org.example.springdb.tiggle.user.model.response;
+
+public class SignupResponse {
+
+}
